+++ react-social.js
@@ -13,8 +13,16 @@
     return !(typeof document === "undefined" || typeof window === "undefined");
   };
 
+  var assign = function(dest, src) {
+    for (var key in src) {
+      dest[key] = src[key];
+    }
+
+    return dest;
+  };
+
   var spread = function (obj, omit) {
-    var clone = React.__spread({}, obj);
+    var clone = assign({}, obj);
 
     omit.forEach(function (key) {
       delete clone[key];
@@ -24,6 +32,8 @@
   };
 
   var jsonp = function (url, cb) {
+    cb = cb || function () { };
+    var called = false;
     var now = +new Date(),
       id = now + "_" + Math.floor(Math.random()*1000);
 
@@ -35,7 +45,21 @@
     script.setAttribute("src", query);
     document.body.appendChild(script);
 
-    window[callback] = cb;
+    setTimeout(function () {
+      if (!called) {
+        called = true;
+        cb(new Error("jsonp timeout"));
+      }
+    }, 10000);
+
+    window[callback] = function () {
+      var args = Array.prototype.slice.call(arguments, 0);
+      args.unshift(null);
+      if (!called) {
+        called = true;
+        cb.apply(this, args);
+      }
+    };
   };
 
 
@@ -75,7 +99,9 @@
   var exports = {};
 
   var Count = {
-    propTypes: {
+    displayName: "Count"
+
+    , propTypes: {
       element: React.PropTypes.string
       , url: React.PropTypes.string
     }
@@ -84,12 +110,13 @@
       var location = "";
 
       if (isBrowser()) {
-        location = window.location;
+        location = window.location.href;
       }
 
       return {
         url: location
         , element: "span"
+        , onCount: function () { }
       };
     }
 
@@ -99,7 +126,7 @@
       };
     }
 
-    , componentWillMount: function () {
+    , componentDidMount: function () {
       this.updateCount();
     }
 
@@ -111,6 +138,10 @@
           this.updateCount();
         });
       }
+    }
+
+    , componentDidUpdate: function () {
+      this.props.onCount(this.state.count);
     }
 
     , updateCount: function () {
@@ -126,7 +157,12 @@
 
       var url = this.constructUrl();
 
-      jsonp(url, function (data) {
+      jsonp(url, function (err, data) {
+        if (err) {
+          console.warn("react-social: jsonp timeout for url " + url);
+          return this.setState({count: 0});
+        }
+
         this.setState({
           count: this.extractCount(data)
         });
@@ -147,19 +183,23 @@
   };
 
   var Button = {
-    propTypes: {
+    displayName: "Button"
+
+    , propTypes: {
       element: React.PropTypes.string
       , url: React.PropTypes.string
       , media: React.PropTypes.string
       , message: React.PropTypes.string
       , onClick: React.PropTypes.func
+      , target: React.PropTypes.string
+      , _open: React.PropTypes.bool
     }
 
     , getDefaultProps: function () {
       var location = "";
 
       if (isBrowser()) {
-        location = window.location;
+        location = window.location.href;
       }
 
       return {
@@ -168,22 +208,15 @@
         , media: ""
         , message: ""
         , onClick: function () { }
+        , _open: true
       };
     }
 
     , click: function (e) {
-      this.props.onClick(e);
-      if (isBrowser()) {
-        var result = this.constructUrl();
-
-        if (typeof result === "object") {
-          var url = result[0]
-          var target = result[1]
-        } else {
-          var url = result
-          var target = "_blank"
-        }
-
+      var url = this.constructUrl();
+      var target = this.props.target;
+      this.props.onClick(e, url, target);
+      if (isBrowser() && this.props._open) {
         window.open(url, target);
       }
     }
@@ -193,14 +226,22 @@
 
       return React.createElement(
         this.props.element
-        , React.__spread({ "onClick": this.click }, other)
+        , assign({ "onClick": this.click }, other)
       );
+    }
+  };
+
+  var DefaultBlankTarget = {
+    getDefaultProps: function () {
+      return {target: "_blank"};
     }
   };
 
   /* Counts */
   exports.FacebookCount = React.createClass({
-    mixins: [Count]
+    displayName: "FacebookCount"
+
+    , mixins: [Count]
 
     , constructUrl: function () {
       var fql = encodeURIComponent("select like_count, share_count from link_stat where url = '" + this.props.url + "'")
@@ -216,8 +257,38 @@
     }
   });
 
+  exports.TwitterCount = React.createClass({
+    displayName: "TwitterCount"
+
+    , mixins: [Count]
+
+    , constructUrl: function () {
+      return "https://count.donreach.com/?callback=@&url=" + encodeURIComponent(this.props.url) + "&providers=all";
+    }
+
+    , extractCount: function (data) {
+      return data.shares.twitter || 0;
+    }
+  });
+
+  exports.GooglePlusCount = React.createClass({
+    displayName: "GooglePlusCount"
+
+    , mixins: [Count]
+
+    , constructUrl: function () {
+      return "https://count.donreach.com/?callback=@&url=" + encodeURIComponent(this.props.url) + "&providers=google";
+    }
+
+    , extractCount: function (data) {
+      return data.shares.google || 0;
+    }
+  });
+
   exports.PinterestCount = React.createClass({
-    mixins: [Count]
+    displayName: "PinterestCount"
+
+    , mixins: [Count]
 
     , constructUrl: function () {
       return "https://api.pinterest.com/v1/urls/count.json?callback=@&url="
@@ -229,8 +300,45 @@
     }
   });
 
+  exports.LinkedInCount = React.createClass({
+    displayName: "LinkedInCount"
+
+    , mixins: [Count]
+
+    , constructUrl: function () {
+      return "https://www.linkedin.com/countserv/count/share?url=" + encodeURIComponent(this.props.url) + "&callback=@&format=jsonp";
+    }
+
+    , extractCount: function (data) {
+      return data.count || 0;
+    }
+  });
+
+  exports.RedditCount = React.createClass({
+    displayName: "RedditCount"
+
+    , mixins: [Count]
+
+    , constructUrl: function () {
+      return "https://www.reddit.com/api/info.json?jsonp=@&url=" + encodeURIComponent(this.props.url);
+    }
+
+    , extractCount: function (data) {
+      var count = 0;
+      var chs = data.data.children;
+
+      for (var i = 0; i < chs.length; i++) {
+        count += chs[i].data.score;
+      }
+
+      return count;
+    }
+  });
+
   exports.VKontakteCount = React.createClass({
-    mixins: [Count]
+    displayName: "VKontakteCount"
+
+    , mixins: [Count]
 
     , fetchCount: function (cb) {
       var index = Math.floor(Math.random() * 10000)
@@ -240,13 +348,52 @@
     }
   });
 
+
+  exports.TumblrCount = React.createClass({
+    displayName: "TumblrCount"
+
+    , mixins: [Count]
+
+    , constructUrl: function () {
+      return "http://api.tumblr.com/v2/share/stats?url="
+             + encodeURIComponent(this.props.url) + "&callback=@";
+    }
+
+    , extractCount: function (data) {
+      return data.response.note_count || 0;
+    }
+  });
+
+  exports.PocketCount = React.createClass({
+    displayName: "PocketCount"
+
+    , mixins: [Count]
+
+    , constructUrl: function () {
+      return "https://count.donreach.com/?callback=@&url=" + encodeURIComponent(this.props.url) + "&providers=pocket";
+    }
+
+    , extractCount: function (data) {
+      return data.shares.pocket || 0;
+    }
+  });
+
   /* Buttons */
   exports.FacebookButton = React.createClass({
-    mixins: [Button]
+    displayName: "FacebookButton"
+
+    , mixins: [Button, DefaultBlankTarget]
+
+    , propTypes: {
+      appId: React.PropTypes.oneOfType([
+        React.PropTypes.string,
+        React.PropTypes.number
+      ])
+    }
 
     , constructUrl: function () {
       return "https://www.facebook.com/dialog/feed?"
-             + "app_id=145634995501895"
+             + "app_id=" + encodeURIComponent(this.props.appId)
              + "&display=popup&caption=" + encodeURIComponent(this.props.message)
              + "&link=" + encodeURIComponent(this.props.url)
              + "&redirect_uri=" + encodeURIComponent("https://www.facebook.com/")
@@ -254,7 +401,9 @@
   });
 
   exports.TwitterButton = React.createClass({
-    mixins: [Button]
+    displayName: "TwitterButton"
+
+    , mixins: [Button, DefaultBlankTarget]
 
     , constructUrl: function () {
       var msg = this.props.message === "" ?
@@ -264,19 +413,23 @@
   });
 
   exports.EmailButton = React.createClass({
-    mixins: [Button]
+    displayName: "EmailButton"
+
+    , mixins: [Button]
+
+    , getDefaultProps: function () {
+      return {target: "_self"};
+    }
 
     , constructUrl: function () {
-      return [
-        "mailto:?subject=" + encodeURIComponent(this.props.message)
-             + "&body=" + encodeURIComponent(this.props.url),
-        "_self"
-      ]
+      return "mailto:?subject=" + encodeURIComponent(this.props.message) + "&body=" + encodeURIComponent(this.props.url);
     }
   });
 
   exports.PinterestButton = React.createClass({
-    mixins: [Button]
+    displayName: "PinterestButton"
+
+    , mixins: [Button, DefaultBlankTarget]
 
     , propTypes: {
       media: React.PropTypes.string.isRequired
@@ -285,13 +438,16 @@
     , constructUrl: function () {
       var url = "https://pinterest.com/pin/create/button/?url="
                 + encodeURIComponent(this.props.url) + "&media="
-                + encodeURIComponent(this.props.media);
+                + encodeURIComponent(this.props.media) + "&description="
+                + encodeURIComponent(this.props.message);
       return url;
     }
   });
 
   exports.VKontakteButton = React.createClass({
-      mixins: [Button]
+    displayName: "VKontakteButton"
+
+    , mixins: [Button, DefaultBlankTarget]
 
     , constructUrl: function () {
         return "http://vk.com/share.php?url=" + encodeURIComponent(this.props.url);
@@ -299,7 +455,9 @@
   });
 
   exports.GooglePlusButton = React.createClass({
-      mixins: [Button]
+    displayName: "GooglePlusButton"
+
+    , mixins: [Button, DefaultBlankTarget]
 
     , constructUrl: function () {
         return "https://plus.google.com/share?url=" + encodeURIComponent(this.props.url);
@@ -307,7 +465,9 @@
   });
 
   exports.RedditButton = React.createClass({
-      mixins: [Button]
+    displayName: "RedditButton"
+
+    , mixins: [Button, DefaultBlankTarget]
 
     , constructUrl: function () {
         return "https://www.reddit.com/submit?url=" + encodeURIComponent(this.props.url);
@@ -315,7 +475,9 @@
   });
 
   exports.LinkedInButton = React.createClass({
-    mixins: [Button]
+    displayName: "LinkedInButton"
+
+    , mixins: [Button, DefaultBlankTarget]
 
     , constructUrl: function () {
       return "https://www.linkedin.com/shareArticle?url=" + encodeURIComponent(this.props.url);
@@ -323,7 +485,9 @@
   });
 
   exports.SmsButton = React.createClass({
-    mixins: [Button]
+    displayName: "SmsButton"
+
+    , mixins: [Button]
 
     , constructUrl: function () {
       var body = encodeURIComponent(this.props.message);
@@ -333,6 +497,36 @@
         "sms:?&body=" + body,
         "_self"
       ]
+    }
+  });
+
+  exports.XingButton = React.createClass({
+    displayName: "XingButton"
+
+    , mixins: [Button, DefaultBlankTarget]
+
+    , constructUrl: function () {
+      return "https://www.xing.com/app/user?op=share;url=" + encodeURIComponent(this.props.url) + ";title=" + encodeURIComponent(this.props.message);
+    }
+  });
+
+  exports.TumblrButton = React.createClass({
+    displayName: "TumblrButton"
+
+    , mixins: [Button, DefaultBlankTarget]
+
+    , constructUrl: function () {
+      return "https://www.tumblr.com/widgets/share/tool?posttype=link&title=" + encodeURIComponent(this.props.message) + "&content=" + encodeURIComponent(this.props.url) + "&canonicalUrl=" + encodeURIComponent(this.props.url) + "&shareSource=tumblr_share_button";
+    }
+  });
+
+  exports.PocketButton = React.createClass({
+    displayName: "PocketButton"
+
+    , mixins: [Button, DefaultBlankTarget]
+
+    , constructUrl: function () {
+      return "https://getpocket.com/save?url=" + encodeURIComponent(this.props.url) + "&title=" + encodeURIComponent(this.props.message);
     }
   });