+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+##### 1.4.0 (2016-05-25)
+
+* Add PocketButton, PocketCount and TwitterCount.
+
+##### 1.3.0 (2016-05-20)
+
+* Add TumblrCount, TumblrButton and XingButton.
+
+##### 1.2.0 (2016-05-18)
+
+* Move `updateCount` to `componentDidMount`.
+* Add `target` prop.
+
+##### 1.1.1 (2016-04-28)
+
+* Pinterest now requires a message.
+
+##### 1.1.0 (2016-04-28)
+
+* Add LinkedInCount, GooglePlusCount and RedditCount.
+* Add tests.
+
+##### 1.0.7 (2016-04-19)
+
+* Remove React.__spread.
+
+##### 1.0.6 (2016-03-21)
+
+* Fix FacebookButton app id.
+
 ##### 1.0.0 (2015-11-27)
 
 * Twitter counts have been deprecated: https://blog.twitter.com/2015/hard-decisions-for-a-sustainable-platform